# public/apidocs/api-docs.yaml
swagger: '2.0'
info:
  description: OpenWeatherMap API
  version: 1.0.0
  title: OpenWeatherMap API
  # put the contact info for your development or API team
paths:
  /import_city:
    get:
      tags:
      - import_city
      summary: import city from openWeatherMap
      operationId: import_city
      description: |
        import city from openWeatherMap
      produces:
      - application/json
      parameters:
      - in: query
        name: city_name
        description: city name
        type: string
      responses:
        200:
          description: search results matching criteria
        400:
          description: bad input parameter
...
host: 0.0.0.0:3000
basePath: /api
schemes:
 - http